% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audubon-package.R
\docType{data}
\name{hiroba}
\alias{hiroba}
\title{Whole tokens of 'Porano no Hiroba' written by Miyazawa Kenji
from Aozora Bunko}
\format{
An object of class \code{data.frame} with 26849 rows and 5 columns.
}
\usage{
hiroba
}
\description{
A tidy text data of \code{audubon::polano} that tokenized with 'MeCab'.
}
\examples{
head(hiroba)
}
\keyword{datasets}
