% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.R
\name{invert_kana}
\alias{invert_kana}
\title{Invert Kana and Roman vice versa}
\usage{
invert_kana(str)
}
\arguments{
\item{str}{Character vector.}
}
\value{
Character vector.
}
\description{
Invert Kana and Roman vice versa
}
