% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.R
\name{invert_to_kana}
\alias{invert_to_kana}
\title{Invert Roman to Hiragana}
\usage{
invert_to_kana(chr)
}
\arguments{
\item{chr}{Character vector.}
}
\value{
A named character vector.
}
\description{
Invert Roman to Hiragana
}
\details{
It is much stricter than the character class translation of ICU
such that this inversion fails and returns an \code{NA_character}
when the element of vector contains any non-Latin strings.
}
