% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.R
\name{invert_to_roman}
\alias{invert_to_roman}
\title{Invert Hiragana to Roman}
\usage{
invert_to_roman(chr)
}
\arguments{
\item{chr}{Character vector.}
}
\value{
A named character vector.
}
\description{
Invert Hiragana to Roman
}
\details{
It is much stricter than the character class translation of ICU
such that this inversion fails and returns an \code{NA_character}
when the element of vector contains any non-Hiragana strings.
}
