% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{ngram_tokenizer}
\alias{ngram_tokenizer}
\title{Ngrams tokenizer}
\usage{
ngram_tokenizer(n = 1L, skip_word_none = FALSE, locale = NULL)
}
\arguments{
\item{n}{Integer.}

\item{skip_word_none}{Logical.}

\item{locale}{Character scalar. Pass a \code{NULL} or an empty string for default locale.}
}
\value{
ngram tokenizer function
}
\description{
Make an ngram tokenizer function.
}
