% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audubon-package.R
\docType{data}
\name{polano}
\alias{polano}
\title{Whole text of 'Porano no Hiroba' written by Miyazawa Kenji
from Aozora Bunko}
\format{
An object of class \code{character} of length 899.
}
\source{
\url{https://www.aozora.gr.jp/cards/000081/files/1935_ruby_19924.zip}
}
\usage{
polano
}
\description{
Whole text of 'Porano no Hiroba' written by Miyazawa Kenji
from Aozora Bunko
}
\examples{
head(polano)
}
\keyword{datasets}
