% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strj-rewrite.R
\name{read_rewrite_def}
\alias{read_rewrite_def}
\title{Read rewrite.def file}
\usage{
read_rewrite_def(def_path = system.file("rewrite.def", package = "audubon"))
}
\arguments{
\item{def_path}{Character scalar; path to the rewriting definition file.}
}
\value{
List.
}
\description{
Read rewrite.def file
}
