% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strj-fill.R
\name{strj_fill_iter_mark}
\alias{strj_fill_iter_mark}
\title{Fill Japanese iteration marks}
\usage{
strj_fill_iter_mark(text)
}
\arguments{
\item{text}{Character vector.}
}
\value{
Character vector.
}
\description{
Fill Japanese iteration marks (Odoriji) with their previous characters.
}
