% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strj-japanese.R
\name{strj_hiranganize}
\alias{strj_hiranganize}
\title{Hiraganize Japanese characters}
\usage{
strj_hiranganize(text)
}
\arguments{
\item{text}{Character vector.}
}
\value{
Character vector.
}
\description{
Hiraganize Japanese characters
}
