% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strj-invert.R
\name{strj_invert_to_kana}
\alias{strj_invert_to_kana}
\title{Invert Roman to Hiragana}
\usage{
strj_invert_to_kana(text)
}
\arguments{
\item{text}{Character vector.}
}
\value{
A named character vector.
}
\description{
Invert Japanese Roman to Hiragana
following JIS X 4063:2000 (deprecated in 2010/01/20).
}
\details{
It is much stricter than the character class translation of ICU
such that this inversion fails and returns an \code{NA_character_}
when the element of vector contains any non-Latin strings.
}
