% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strj-invert.R
\name{strj_invert_to_roman}
\alias{strj_invert_to_roman}
\title{Invert Hiragana to Roman}
\usage{
strj_invert_to_roman(text)
}
\arguments{
\item{text}{Character vector.}
}
\value{
A named character vector.
}
\description{
Invert Japanese Hiragana to Roman
following JIS X 4063:2000 (deprecated in 2010/01/20).
}
\details{
It is much stricter than the character class translation of ICU
such that this inversion fails and returns an \code{NA_character_}
when the element of vector contains any non-Hiragana strings.
}
\section{Limitations}{

In case you use this function under Windows with Japanese locale,
some romanization would not work porperly due to its default encoding.
}

