% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strj-japanese.R
\name{strj_katakanize}
\alias{strj_katakanize}
\title{Katakanize Japanese characters}
\usage{
strj_katakanize(text)
}
\arguments{
\item{text}{Character vector.}
}
\value{
Character vector.
}
\description{
Katakanize Japanese characters
}
