% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strj-normalize.R
\name{strj_normalize}
\alias{strj_normalize}
\title{Convert text following the rules of 'NEologd'}
\usage{
strj_normalize(text)
}
\arguments{
\item{text}{Character vector.}
}
\value{
Character vector.
}
\description{
Convert characters into normalized style
basing on rules that is recommended by the Neologism dictionary for MeCab.
}
\seealso{
\url{https://github.com/neologd/mecab-ipadic-neologd/wiki/Regexp.ja}
}
