% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strj-japanese.R
\name{strj_romanize}
\alias{strj_romanize}
\title{Romanize Japanese Hiragana and Katakana}
\usage{
strj_romanize(
  text,
  config = c("wikipedia", "traditional hepburn", "modified hepburn", "kunrei", "nihon")
)
}
\arguments{
\item{text}{Character vector.}

\item{config}{Config used to romanize.}
}
\value{
Character vector.
}
\description{
Romanize Japanese Hiragana and Katakana
}
