% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strj-segment.R
\name{strj_segment}
\alias{strj_segment}
\title{Segment text into phrases}
\usage{
strj_segment(text)
}
\arguments{
\item{text}{Character vector.}
}
\value{
List.
}
\description{
Segment Japanese text into several phrases
using the 'google/budoux' JavaScript module.
}
