% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strj-tokenize.R
\name{strj_tokenize}
\alias{strj_tokenize}
\title{Split text into tokens}
\usage{
strj_tokenize(
  text,
  format = c("list", "data.frame"),
  engine = c("stringi", "budoux", "tinyseg"),
  split = FALSE,
  ...
)
}
\arguments{
\item{text}{Character vector to be tokenized.}

\item{format}{Output format. Choose \code{list} or \code{data.frame}.}

\item{engine}{Tokenizer name. One of "stringi", "budoux", "tinyseg".}

\item{split}{Logical. If \code{TRUE}, the function splits vectors}
}
\value{
A list or a data.frame.
}
\description{
Splits text into several tokens using specified tokenizer.
}
\examples{
strj_tokenize(
  paste0(
    "\u3042\u306e\u30a4\u30fc\u30cf\u30c8",
    "\u30fc\u30f4\u30a9\u306e\u3059\u304d",
    "\u3068\u304a\u3063\u305f\u98a8"
  )
)
strj_tokenize(
  paste0(
    "\u3042\u306e\u30a4\u30fc\u30cf\u30c8",
    "\u30fc\u30f4\u30a9\u306e\u3059\u304d",
    "\u3068\u304a\u3063\u305f\u98a8"
  ),
  format = "data.frame"
)
}
