% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strj-japanese.R
\name{strj_transcribe_num}
\alias{strj_transcribe_num}
\title{Transcribe Arabic to Kansuji}
\usage{
strj_transcribe_num(int)
}
\arguments{
\item{int}{Integers.}
}
\value{
Character vector.
}
\description{
Transcribe Arabic integers to Kansuji.
}
