% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audubon-package.R
\docType{package}
\name{audubon-package}
\alias{audubon}
\alias{audubon-package}
\title{audubon:  R Package for Various Japanese Text Processing}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

An R package for various Japanese text processing.
}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

}
\keyword{internal}
