% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audubon-package.R
\docType{data}
\name{polano}
\alias{polano}
\title{Whole text of 'Porano no Hiroba' written by Miyazawa Kenji
from Aozora Bunko}
\format{
An object of class \code{character} of length 899.
}
\source{
\url{https://www.aozora.gr.jp/cards/000081/files/1935_ruby_19924.zip}
}
\usage{
polano
}
\description{
Whole text of 'Porano no Hiroba' written by Miyazawa Kenji
from Aozora Bunko
}
\details{
A dataset containing the text of Miyazawa Kenji's novel
"Porano no Hiroba" which was published in 1934, the year after Kenji's death.
Copyright of this work has expired since more than 70 years have passed after the author's death.

The UTF-8 plain text is sourced from \url{https://www.aozora.gr.jp/cards/000081/card1935.html} and is cleaned of meta data.
}
\examples{
head(polano)
}
\keyword{datasets}
