% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strj-rewrite.R
\name{read_rewrite_def}
\alias{read_rewrite_def}
\title{Read a rewrite.def file}
\usage{
read_rewrite_def(
  def_path = system.file("def/rewrite.def", package = "audubon")
)
}
\arguments{
\item{def_path}{Character scalar; path to the rewriting definition file.}
}
\value{
A list.
}
\description{
Read a rewrite.def file
}
\examples{
str(read_rewrite_def())
}
