% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strj-fill.R
\name{strj_fill_iter_mark}
\alias{strj_fill_iter_mark}
\title{Fill Japanese iteration marks}
\usage{
strj_fill_iter_mark(text)
}
\arguments{
\item{text}{Character vector.}
}
\value{
A character vector.
}
\description{
Fills Japanese iteration marks (Odori-ji) with their previous characters
if the element has more than 5 characters.
}
\examples{
strj_fill_iter_mark(c(
  "\u3042\u3044\u3046\u309d\u3003\u304b\u304d",
  "\u91d1\u5b50\u307f\u3059\u309e",
  "\u306e\u305f\u308a\u3033\u3035\u304b\u306a",
  "\u3057\u308d\uff0f\u2033\uff3c\u3068\u3057\u305f"
))
}
