% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strj-rewrite.R
\name{strj_rewrite_as_def}
\alias{strj_rewrite_as_def}
\title{Rewrite text using rewrite.def}
\usage{
strj_rewrite_as_def(text, as = read_rewrite_def())
}
\arguments{
\item{text}{Character vector.}

\item{as}{List.}
}
\value{
Character vector
}
\description{
Rewrite text using rewrite.def
}
