% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strj-rewrite.R
\name{strj_rewrite_as_def}
\alias{strj_rewrite_as_def}
\title{Rewrite text using rewrite.def}
\usage{
strj_rewrite_as_def(text, as = read_rewrite_def())
}
\arguments{
\item{text}{Character vector to be normalized.}

\item{as}{List.}
}
\value{
A character vector.
}
\description{
Rewrites text using a 'rewrite.def' file.
}
\examples{
strj_rewrite_as_def(
  paste0(
    "\u2015\u2015\u5357\u30a2\u30eb",
    "\u30d7\u30b9\u306e\u3000\u5929",
    "\u7136\u6c34-\u3000\uff33\uff50",
    "\uff41\uff52\uff4b\uff49\uff4e\uff47*",
    "\u3000\uff2c\uff45\uff4d\uff4f\uff4e+",
    "\u3000\u30ec\u30e2\u30f3\u4e00\u7d5e\u308a"
  )
)
strj_rewrite_as_def(
  "\u60e1\u3068\u5047\u9762\u306e\u30eb\u30fc\u30eb",
  read_rewrite_def(system.file("def/kyuji.def", package = "audubon"))
)
}
