% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strj-japanese.R
\name{strj_romanize}
\alias{strj_romanize}
\title{Romanize Japanese Hiragana and Katakana}
\usage{
strj_romanize(
  text,
  config = c("wikipedia", "traditional hepburn", "modified hepburn", "kunrei", "nihon")
)
}
\arguments{
\item{text}{Character vector.}

\item{config}{Config used to romanize.}
}
\value{
Character vector.
}
\description{
Romanize Japanese Hiragana and Katakana
}
\examples{
strj_romanize("\u3042\u306e\u30a4\u30fc\u30cf\u30c8\u30fc\u30f4\u30a9\u306e\u3059\u304d\u3068\u304a\u3063\u305f\u98a8")
}
