% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strj-tokenize.R
\name{strj_tokenize}
\alias{strj_tokenize}
\title{Simply tokenize sentence}
\usage{
strj_tokenize(sentence)
}
\arguments{
\item{sentence}{Character vector to be tokenized.}
}
\value{
data.frame.
}
\description{
Split the given sentence into tokens
using \code{stringi::stri_split_boundaries(type = "word")}.
}
