% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strj-japanese.R
\name{strj_transcribe_num}
\alias{strj_transcribe_num}
\title{Transcribe Arabic to Kansuji}
\usage{
strj_transcribe_num(int)
}
\arguments{
\item{int}{Integers.}
}
\value{
A character vector.
}
\description{
Transcribes Arabic integers to Kansuji with auxiliary numerals.
}
\details{
As its implementation is limited, this function can only transcribe
numbers up to trillions.
In case you convert much bigger numbers, try to use the 'arabic2kansuji' package.
}
\examples{
strj_transcribe_num(c(10L, 31415L))
}
