% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strj-japanese.R
\name{strj_transcribe_num}
\alias{strj_transcribe_num}
\title{Transcribe integers into Japanese kanji numerals}
\usage{
strj_transcribe_num(int)
}
\arguments{
\item{int}{An integer vector to transcribe.}
}
\value{
A character vector containing kanji numeral representations.
}
\description{
Converts integer values to their Japanese kanji numeral representations.

This function transcribes integers up to the trillions place into kanji
numerals. For larger numbers or more comprehensive numeral support, consider
using the CRAN package arabic2kansuji.
}
\examples{
strj_transcribe_num(c(10L, 31415L))
}
